add auth module

+++ laravel-api/app/Http/Controllers/Api/ProductsCartController.php
@@ -2,48 +2,51 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Cart;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use App\Cart;
 
 class ProductsCartController extends Controller
 {
     public function index()
     {
         return Cart::with('product')
-                ->orderBy('created_at', 'desc')
-                ->get();
+            ->where('user_id', Auth::user()->id)
+            ->orderBy('created_at', 'desc')
+            ->get();
     }
 
     public function store(Request $request)
     {
-        $item = Cart::where('product_id', $request->product_id);
+        $item = Cart::where('user_id', Auth::user()->id)->where('product_id', $request->product_id);
 
         if ($item->count()) {
-            $item->increment('quantity');
+            $item->increment('quantity', $request->quantity);
             $item = $item->first();
         } else {
             $item = Cart::forceCreate([
+                'user_id' => Auth::user()->id,
                 'product_id' => $request->product_id,
-                'quantity' => 1,
+                'quantity' => $request->quantity,
             ]);
         }
 
         return response()->json([
             'quantity' => $item->quantity,
-            'product' => $item->product
+            'product' => $item->product,
         ]);
     }
 
     public function destroy($productId)
     {
-    	$item = Cart::where('product_id', $productId)->delete();
+        $item = Cart::where('user_id', Auth::user()->id)->where('product_id', $productId)->delete();
 
         return response(null, 200);
     }
 
     public function destroyAll()
     {
-        Cart::truncate();
+        Cart::where('user_id', Auth::user()->id)->delete();
     }
 }
